+++ lib/src/actions/base_actions.dart
@@ -5,4 +5,5 @@ part 'actions/explosion.dart';
 part 'actions/hands.dart';
 part 'actions/immortality.dart';
 part 'actions/points.dart';
+part 'actions/region.dart';
 part 'actions/transfer.dart';